% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{mult.aov}
\alias{mult.aov}
\title{Multiple Group Anova.}
\usage{
mult.aov(data, group1, group2, value, level)
}
\value{
Return a datafram
}
\description{
\code{mult.aov} Multiple Group Anova.
}
\author{
Xiang LI <lixiang117423@foxmail.com>
}
