% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{multGroupTtest}
\alias{multGroupTtest}
\title{Multiple Group Student's-test.}
\usage{
multGroupTtest(data, group1, group2, CK, value, level)
}
\value{
Return a datafram
}
\description{
\code{multGroupTtest} Multiple Group Student's-test.
}
\author{
Xiang LI <lixiang117423@foxmail.com>
}
